cli, server: store available ram and CPU cores in node summaries
Includes moving the GetTotalMemory function from the server package to
the server/status package, so it can be used by the stats recorder and
also the base Server where it was originally used.

Release note: None
